moved profile function from GameController to ProfileController
