+++ lib/controllers/index.js
@@ -10,6 +10,7 @@ const WeblauncherController = require("./WeblauncherController");
 const RaidController = require("./RaidController");
 const ItemController = require("./ItemController");
 const HideoutController = require("./HideoutController");
+const ProfileController = require("./ProfileController");
 
 module.exports = {
     ...AccountController,
@@ -23,5 +24,6 @@ module.exports = {
     ...WeblauncherController,
     ...RaidController,
     ...ItemController,
-    ...HideoutController
+    ...HideoutController,
+    ...ProfileController
 };
